src:Redactor: move SBOM handling to library code
Same rationale as database backend. We can re-use this code where
necessary, and simplifies the script, which should exclusively focused
on presenting this functionality to users.
