+++ privateSBOMExchange/src/petra/lib/sbom/__init__.py
@@ -1,0 +1,1 @@
+from .spdx import *
